[riscv64] Keep the type consistent in in_range
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/3516829
Auto-Submit: Yahan Lu <yahan@iscas.ac.cn>
Commit-Queue: Yahan Lu <yahan@iscas.ac.cn>
Cr-Commit-Position: refs/heads/main@{#79451}

+++ src/codegen/riscv64/register-riscv64.h
@@ -300,7 +300,7 @@ class FPURegister : public RegisterBase<FPURegister, kDoubleAfterLast> {
   // this cl, in order to facilitate modification, it is assumed that the vector
   // register and floating point register are shared.
   VRegister toV() const {
-    DCHECK(base::IsInRange(code(), 0, kVRAfterLast - 1));
+    DCHECK(base::IsInRange(static_cast<int>(code()), 0, kVRAfterLast - 1));
     return VRegister(code());
   }
